[#3] MainState 타입 분리
