+++ src/main.rs
@@ -3,12 +3,14 @@ mod custom_style;
 mod dialog;
 mod file;
 mod modal;
+mod state;
 mod static_assets;
 
 use custom_style::SettingButtonStyle;
 use iced::widget::{self, button, column, container, text, Column};
 use iced::{alignment, theme, Color, Element, Length, Sandbox, Settings, Size, Theme};
 use modal::Modal;
+use state::{MainState, MusicList};
 
 pub fn main() -> iced::Result {
     let app_data_path = config::get_app_data_path();
@@ -24,107 +26,8 @@ pub fn main() -> iced::Result {
     Player::run(setting)
 }
 
-#[derive(Debug, Clone, Default)]
-pub struct Music {
-    title: String,
-    pub file_path: String,
-}
-
-#[derive(Debug, Clone)]
-pub struct MusicList {
-    list: Vec<Music>,
-}
-
-impl Default for MusicList {
-    fn default() -> Self {
-        Self {
-            list: vec![
-                Music {
-                    title: "test1".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test2".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test3".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test4".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test5".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test6".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test7".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test8".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test9".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test10".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test11".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test12".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test13".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test14".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test10".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test11".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test12".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test13".into(),
-                    file_path: "test".into(),
-                },
-                Music {
-                    title: "test14".into(),
-                    file_path: "test".into(),
-                },
-            ],
-        }
-    }
-}
-
 pub struct Player {
-    title: String,
-    value: i32,
-    music_list: MusicList,
-    on_play: bool,
+    main_state: MainState,
     show_setting_modal: bool,
 }
 
@@ -142,10 +45,12 @@ impl Sandbox for Player {
 
     fn new() -> Self {
         Self {
-            title: "test name".into(),
-            value: 0,
-            music_list: MusicList::default(),
-            on_play: false,
+            main_state: MainState {
+                title: "test name".into(),
+                value: 0,
+                music_list: MusicList::default(),
+                on_play: false,
+            },
             show_setting_modal: false,
         }
     }
@@ -161,13 +66,13 @@ impl Sandbox for Player {
     fn update(&mut self, message: PlayerMessage) {
         match message {
             PlayerMessage::ResumeOrPausePressed => {
-                self.on_play = !self.on_play;
+                self.main_state.on_play = !self.main_state.on_play;
             }
             PlayerMessage::NextPressed => {
-                self.value += 1;
+                self.main_state.value += 1;
             }
             PlayerMessage::PreviousPressed => {
-                self.value -= 1;
+                self.main_state.value -= 1;
             }
             PlayerMessage::OpenSettingModal => {
                 self.show_setting_modal = true;
@@ -184,7 +89,7 @@ impl Sandbox for Player {
                 container(
                     container(column!(
                         container(self.setting_button()).padding(0),
-                        container(text(self.title.as_str()).size(15))
+                        container(text(self.main_state.title.as_str()).size(15))
                             .padding(10)
                             .align_x(alignment::Horizontal::Center)
                             .width(Length::Fill),
@@ -268,7 +173,7 @@ impl Player {
             .align_items(iced::Alignment::Center)
             .width(Length::Fill);
 
-        for value in self.music_list.list.iter() {
+        for value in self.main_state.music_list.list.iter() {
             column = column.push(text(value.title.as_str()));
         }
 
@@ -296,7 +201,7 @@ impl Player {
         .width(Length::Fixed(50_f32))
         .height(Length::Fixed(50_f32));
 
-        let resume_or_pause_button_text = if self.on_play { "||" } else { ">>" };
+        let resume_or_pause_button_text = if self.main_state.on_play { "||" } else { ">>" };
 
         let resume_or_pause_button = button(
             text(resume_or_pause_button_text)
